fix:improve type structure
