+++ v2/pink-sb/src/lib/expandable-table/index.ts
@@ -2,16 +2,7 @@ import Root from './root/Root.svelte';
 import Row from './row/Row.svelte';
 import Cell from './cell/Cell.svelte';
 
-export type RootProp = {
-    single: boolean;
-    openIds: string[];
-    isOpen: (id: string) => boolean;
-    open: (id: string) => void;
-    close: (id: string) => void;
-    toggle: (id: string) => void;
-    register: (id: string) => void;
-    unregister: (id: string) => void;
-};
+export type { ExpandableTableColumn, RootProp } from './types.js';
 
 export default {
     Root,
